src/requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
bcrypt==3.1.4
billiard==3.6.0.0
celery==4.3.0
redis==3.3.8
cffi==1.11.2
click==6.6
cryptography==2.1.4
decorator==4.0.10
dj-database-url==0.4.2
Django==1.11.24
django-appconf==1.0.2
django-compressor==2.1.1
django-cors-headers==2.1.0
django-debug-toolbar==1.6
django-filter==0.14.0
django-widget-tweaks==1.4.1
djangorestframework==3.7.3
Fabric3==1.13.1.post1
gunicorn==19.6.0
html5lib==0.9999999
idna==2.6
kombu==4.6.4
Markdown==2.6.6
olefile==0.44
paramiko==2.4.0
Pillow==4.2.1
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
PyNaCl==1.2.1
pytz==2017.3
ratelim==0.1.6
rcssmin==1.0.6
requests==2.10.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
vine==1.3.0
